EDG-722: Move the Nevsky framework to the protocols.v2 package
The Nevsky framework package moves from com.hivemq.protocols2 to
com.hivemq.protocols.v2, mirroring the SDK's api.v2 rename. References to the
renamed SDK v2 types are updated (api.v2 imports, Tag).

ReuseNoForkTest is rewritten to drop the obsolete '2'-suffix (N1) check, sync
the no-fork name set to the de-suffixed v2 contracts, and walk the protocols.v2
source tree; N2 (no abbreviations) is unchanged.

compileTestJava and spotlessCheck are green; the protocols.v2 tests pass.
